requirements.txt
Django>=5.2
requests>=2.32.0
beautifulsoup4>=4.14.0
html5lib>=1.1
python-dotenv>=1.0.0
dj-database-url>=3.0.0
celery>=5.5.3
redis>=4.5.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
flower>=2.0.0
responses>=0.25.8